resolve #77
- replace `NumberPickerDialogFragment` with `NumberPickerDialog`

- replace `AlertDialogFragment` with `AlertDialogFragment2`
