+++ app/src/main/java/com/suihan74/satena/scenes/entries2/EntriesActivity.kt
@@ -30,7 +30,7 @@ import com.suihan74.satena.NetworkReceiver
 import com.suihan74.satena.R
 import com.suihan74.satena.SatenaApplication
 import com.suihan74.satena.databinding.ActivityEntries2Binding
-import com.suihan74.satena.dialogs.AlertDialogFragment2
+import com.suihan74.satena.dialogs.AlertDialogFragment
 import com.suihan74.satena.dialogs.ReleaseNotesDialogFragment
 import com.suihan74.satena.models.Category
 import com.suihan74.satena.models.PreferenceKey
@@ -476,7 +476,7 @@ class EntriesActivity : AppCompatActivity() {
     /** アクティビティ終了時確認 */
     override fun finish() {
         if (viewModel.isTerminationDialogEnabled && supportFragmentManager.backStackEntryCount <= 1) {
-            val dialog = AlertDialogFragment2.Builder()
+            val dialog = AlertDialogFragment.Builder()
                 .setTitle(R.string.confirm_dialog_title_simple)
                 .setMessage(R.string.app_termination_dialog_msg)
                 .setNegativeButton(R.string.dialog_cancel)
