+++ app/src/main/java/com/suihan74/satena/scenes/preferences/pages/PreferencesEntriesFragment.kt
@@ -12,8 +12,8 @@ import androidx.databinding.InverseMethod
 import com.suihan74.hatenaLib.HatenaClient
 import com.suihan74.satena.R
 import com.suihan74.satena.databinding.FragmentPreferencesEntriesBinding
-import com.suihan74.satena.dialogs.AlertDialogFragment2
-import com.suihan74.satena.dialogs.NumberPickerDialogFragment
+import com.suihan74.satena.dialogs.AlertDialogFragment
+import com.suihan74.satena.dialogs.NumberPickerDialog
 import com.suihan74.satena.models.*
 import com.suihan74.satena.scenes.entries2.CategoriesMode
 import com.suihan74.satena.scenes.entries2.ExtraBottomItemsAlignment
@@ -23,24 +23,9 @@ import com.suihan74.utilities.provideViewModel
 import com.suihan74.utilities.showAllowingStateLoss
 import kotlinx.android.synthetic.main.fragment_preferences_entries.view.*
 
-class PreferencesEntriesFragment :
-    PreferencesFragmentBase(),
-    NumberPickerDialogFragment.Listener
-{
+class PreferencesEntriesFragment : PreferencesFragmentBase() {
     companion object {
         fun createInstance() = PreferencesEntriesFragment()
-
-        private const val DIALOG_BOTTOM_BAR_ITEM_SETTER = "DIALOG_BOTTOM_BAR_ITEM_SETTER"
-        private const val DIALOG_SINGLE_TAP_ACTION = "DIALOG_SINGLE_TAP_ACTION"
-        private const val DIALOG_MULTIPLE_TAP_ACTION = "DIALOG_MULTIPLE_TAP_ACTION"
-        private const val DIALOG_LONG_TAP_ACTION = "DIALOG_LONG_TAP_ACTION"
-        private const val DIALOG_MULTIPLE_TAP_DURATION = "DIALOG_MULTIPLE_TAP_DURATION"
-        private const val DIALOG_HOME_CATEGORY = "DIALOG_HOME_CATEGORY"
-        private const val DIALOG_HOME_TAB = "DIALOG_HOME_TAB"
-        private const val DIALOG_HISTORY_MAX_SIZE_PICKER = "DIALOG_HISTORY_MAX_SIZE_PICKER"
-        private const val DIALOG_ENTRY_READ_ACTION_TYPE = "DIALOG_ENTRY_READ_ACTION_TYPE"
-        private const val DIALOG_CATEGORIES_MODE = "DIALOG_CATEGORIES_MODE"
-        private const val DIALOG_ADDITIONAL_BOTTOM_ITEMS_ALIGNMENT = "DIALOG_ADDITIONAL_BOTTOM_ITEMS_ALIGNMENT"
     }
 
     val viewModel by lazy {
@@ -69,14 +54,13 @@ class PreferencesEntriesFragment :
         view.bottom_bar_item_setter.setOnMenuItemClickListener { args ->
             viewModel.showBottomBarItemSetterDialog(
                 args,
-                childFragmentManager,
-                DIALOG_BOTTOM_BAR_ITEM_SETTER
+                childFragmentManager
             )
         }
 
         // シングルタップ時の動作
         view.preferences_entries_single_tap_action.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_single_tap_action_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(tapActions, viewModel.singleTapAction.value!!.ordinal) { _, which ->
@@ -84,12 +68,12 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_SINGLE_TAP_ACTION)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // 複数回タップ時の動作
         view.preferences_entries_multiple_tap_action.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_multiple_tap_action_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(tapActions, viewModel.multipleTapAction.value!!.ordinal) { _, which ->
@@ -97,12 +81,12 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_MULTIPLE_TAP_ACTION)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // ロングタップ時の動作
         view.preferences_entries_long_tap_action.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_long_tap_action_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(tapActions, viewModel.longTapAction.value!!.ordinal) { _, which ->
@@ -110,17 +94,20 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_LONG_TAP_ACTION)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         view.preferences_entries_multiple_tap_duration.setOnClickListener {
-            NumberPickerDialogFragment.Builder()
-                .setTitle(R.string.pref_entries_multiple_tap_duration_desc)
-                .setMessage(R.string.pref_entries_multiple_tap_duration_dialog_message)
-                .setMinValue(0)
-                .setMaxValue(500)
-                .setDefaultValue(viewModel.multipleTapDuration.value!!.toInt())
-                .showAllowingStateLoss(childFragmentManager, DIALOG_MULTIPLE_TAP_DURATION)
+            val dialog = NumberPickerDialog.createInstance(
+                min = 0,
+                max = 500,
+                default = viewModel.multipleTapDuration.value!!.toInt(),
+                titleId = R.string.pref_entries_multiple_tap_duration_desc,
+                messageId = R.string.pref_entries_multiple_tap_duration_dialog_message
+            ) { value ->
+                viewModel.multipleTapDuration.value = value.toLong()
+            }
+            dialog.showAllowingStateLoss(childFragmentManager)
         }
 
         // ホームカテゴリ
@@ -130,7 +117,7 @@ class PreferencesEntriesFragment :
                 else Category.valuesWithoutSignedIn()
             ).filter { it.willBeHome }
 
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_home_category_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(
@@ -141,14 +128,14 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_HOME_CATEGORY)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // 最初に表示するタブ
         view.preferences_entries_initial_tab.setOnClickListener {
             val items = getTabTitleIds(viewModel.homeCategory.value!!)
 
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_initial_tab_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(
@@ -159,23 +146,26 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_HOME_TAB)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // ブクマ閲覧履歴の最大保存数
         view.preferences_entries_history_max_size.setOnClickListener {
-            NumberPickerDialogFragment.Builder()
-                .setTitle(R.string.pref_entries_history_max_size_dialog_title)
-                .setMessage(R.string.pref_entries_history_max_size_dialog_msg)
-                .setMinValue(1)
-                .setMaxValue(100)
-                .setDefaultValue(viewModel.historyMaxSize.value!!)
-                .showAllowingStateLoss(childFragmentManager, DIALOG_HISTORY_MAX_SIZE_PICKER)
+            val dialog = NumberPickerDialog.createInstance(
+                min = 1,
+                max = 100,
+                default = viewModel.historyMaxSize.value!!,
+                titleId = R.string.pref_entries_history_max_size_dialog_title,
+                messageId = R.string.pref_entries_history_max_size_dialog_msg
+            ) { value ->
+                viewModel.multipleTapDuration.value = value.toLong()
+            }
+            dialog.showAllowingStateLoss(childFragmentManager)
         }
 
         // 「あとで読む」エントリを「読んだ」したときの挙動
         view.preferences_entries_read_action_type.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_read_action_type_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(
@@ -186,12 +176,12 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_ENTRY_READ_ACTION_TYPE)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // カテゴリリストの表示形式
         view.preferences_entries_categories_mode.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_entries_categories_mode_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(
@@ -202,12 +192,12 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_CATEGORIES_MODE)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         // ボトムバーの追加項目の配置方法
         view.extra_bottom_items_alignment_button.setOnClickListener {
-            AlertDialogFragment2.Builder()
+            AlertDialogFragment.Builder()
                 .setTitle(R.string.pref_extra_bottom_items_alignment_desc)
                 .setNegativeButton(R.string.dialog_cancel)
                 .setSingleChoiceItems(
@@ -218,23 +208,10 @@ class PreferencesEntriesFragment :
                 }
                 .dismissOnClickItem(true)
                 .create()
-                .showAllowingStateLoss(childFragmentManager, DIALOG_ADDITIONAL_BOTTOM_ITEMS_ALIGNMENT)
+                .showAllowingStateLoss(childFragmentManager)
         }
 
         return view
-    }
-
-    /** NumberPickerの処理完了 */
-    override fun onCompleteNumberPicker(value: Int, dialog: NumberPickerDialogFragment) {
-        when (dialog.tag) {
-            DIALOG_MULTIPLE_TAP_DURATION -> {
-                viewModel.multipleTapDuration.value = value.toLong()
-            }
-
-            DIALOG_HISTORY_MAX_SIZE_PICKER -> {
-                viewModel.historyMaxSize.value = value
-            }
-        }
     }
 }
